First pass at a JNI-C++ Object Mapping
This code adds a map of shared_ptr to Java.  Instead of using reinterpret_cast to cast the object back and forth between Java and C++, this code puts it into a map and thereby keeps a pointer around in C++ (in the map).  The map was used here for Options, DBOptions, ColumnFamilyOptions, Read/WriteOptions.

This is a work in progress.  Not all Java tests pass.  It is not clear a shared_ptr is the best container (as opposed to a custom one) because some objects are "owned" by the map and some are "owned" by other things (like CompressionOptions or Transaction::WriteOptions).